django_learn/requirements.txt
django

# for django.contrib.admindocs
docutils

# Django community extensions
# Ninja
django-ninja
django-ninja-extra

# DRF
djangorestframework

# development tools
django-debug-toolbar
django-extensions
